+++ apps/api-service/src/intelligence/collections/collections.repository.ts
@@ -1,16 +1,16 @@
 import { Injectable } from "@nestjs/common"
 import { InjectModel } from "@nestjs/mongoose"
 import { DbConnectionMap } from "src/shared/utils/db-connection.map"
-import { Favourite } from "./schemas/favourites.schema"
+import { Collection } from "./schemas/collections.schema"
 import { Model } from "mongoose"
 import { EntityRepository } from "@/shared/entity/entity.repository"
 
 @Injectable()
-export class FavouritesRepository extends EntityRepository<Favourite> {
+export class CollectionsRepository extends EntityRepository<Collection> {
   constructor(
-    @InjectModel(Favourite.name, DbConnectionMap.Intelligence)
-    private favouriteModel: Model<Favourite>
+    @InjectModel(Collection.name, DbConnectionMap.Intelligence)
+    private collectionModel: Model<Collection>
   ) {
-    super(favouriteModel)
+    super(collectionModel)
   }
 }