+++ apps/api-service/src/intelligence/collections/commands/handler/remove-from-collection.handler.ts
@@ -1,0 +1,19 @@
+import { CommandHandler, ICommandHandler } from "@nestjs/cqrs"
+import { CollectionsRepository } from "../../collections.repository"
+import { RemoveCollectionCommand } from "../impl/remove-from-collection.command"
+import objectId from "@/shared/utils/convert-objectid"
+
+@CommandHandler(RemoveCollectionCommand)
+export class RemoveCollectionCommandHandler
+  implements ICommandHandler<RemoveCollectionCommand>
+{
+  constructor(private readonly repository: CollectionsRepository) {}
+
+  async execute(command: RemoveCollectionCommand) {
+    const { userId, modelId } = command
+    return await this.repository.delete({
+      userId: objectId(userId),
+      derivedModel: objectId(modelId),
+    })
+  }
+}