+++ apps/api-service/src/intelligence/collections/commands/impl/add-to-collection.command.ts
@@ -1,4 +1,4 @@
-export class CreateFavouriteCommand {
+export class CreateCollectionCommand {
   constructor(
     public readonly userId: string,
     public readonly modelId: string