+++ apps/api-service/src/intelligence/collections/commands/impl/remove-from-collection.command.ts
@@ -1,4 +1,4 @@
-export class RemoveFavouriteCommand {
+export class RemoveCollectionCommand {
   constructor(
     public readonly userId: string,
     public readonly modelId: string