+++ apps/api-service/src/intelligence/collections/queries/impl/find-collection-list.query.ts
@@ -1,3 +1,3 @@
-export class FindAllFavouritesQuery {
+export class FindAllCollectionsQuery {
   constructor(public readonly userId: string) {}
 }