+++ apps/api-service/src/intelligence/collections/queries/impl/find-if-collected.query.ts
@@ -1,4 +1,4 @@
-export class FindIfFavouritedQuery {
+export class FindIfCollectedQuery {
   constructor(
     public readonly userId: string,
     public readonly modelId: string